design improvements
