--- packages/web/src/components/SpacePage/ConferenceView/MediaGridView/useMediaGridView.ts
@@ -1,9 +1,0 @@
-import { useRemotePeers } from "../../hooks/useRemotePeers";
-import { useLocalUser } from "../../hooks/useLocalUser";
-
-export const useMediaGridView = () => {
-  const localUser = useLocalUser();
-  const remotePeers = useRemotePeers();
-
-  return { mainUser: localUser, otherUsers: remotePeers };
-};
